# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FAF Backend API",
    "version": "1.0.0",
    "description": "API Documentation for Freelance Application Framework",
    "contact": {
      "name": "FAF Developer"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/admin/jobs/pending": {
      "get": {
        "summary": "List pending jobs",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending jobs"
          }
        }
      }
    },
    "/api/admin/jobs/{id}/approve": {
      "put": {
        "summary": "Approve a job",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job approved"
          }
        }
      }
    },
    "/api/admin/jobs/{id}/reject": {
      "put": {
        "summary": "Reject a job",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job rejected"
          }
        }
      }
    },
...
    "/api/contracts/active/me": {
      "get": {
        "summary": "Get worker's active contract",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Active contract details"
          }
        }
      }
    },
    "/api/contracts/my": {
      "get": {
        "summary": "Get all my contracts",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of contracts"
          }
        }
      }
    },
    "/api/contracts/job/{jobId}": {
      "get": {
        "summary": "Get contract by Job ID",
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract details"
          }
        }
      }
    },
    "/api/contracts/{id}": {
      "get": {
        "summary": "Get contract details",
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract details"
          }
        }
      }
    },
    "/api/contracts/{id}/sign": {
      "post": {
        "summary": "Sign a contract",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract signed"
          }
        }
      }
    },
    "/api/contracts/{id}/terminate": {
      "put": {
        "summary": "Terminate a contract",
        "tags": [
          "Contracts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract terminated"
          }
        }
      }
    },
...
    "/api/jobs": {
      "post": {
        "summary": "Create a new job",
        "tags": [
          "Jobs"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "budget",
                  "jobType",
                  "categoryId",
                  "checkpoints"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "jobType": {
                    "type": "string",
                    "enum": [
                      "SHORT_TERM",
                      "LONG_TERM"
                    ]
                  },
                  "budget": {
                    "type": "number"
                  },
                  "categoryId": {
                    "type": "integer"
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date"
                  },
                  "contractContent": {
                    "type": "string"
                  },
                  "checkpoints": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "title",
                        "amount"
                      ],
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "amount": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        },
                        "due_date": {
                          "type": "string",
                          "format": "date"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List all jobs",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs"
          }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "summary": "Get job details",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job details"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
...
    "/api/reviews/contract/{contractId}": {
      "get": {
        "summary": "Get reviews for a specific contract",
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contractId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of reviews"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "summary": "List all users (Admin only)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of users"
          }
        }
      }
    },
...
  },
...
}